refactor: move injectors under pkg/injector
